+++ example/src/DependencyInjection/NetzmachtContaoWorkflowExampleExtension.php
@@ -20,10 +20,13 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
 use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 
+/**
+ * Class NetzmachtContaoWorkflowExampleExtension
+ */
 final class NetzmachtContaoWorkflowExampleExtension extends Extension
 {
     /**
-     * @inheritDoc
+     * {@inheritDoc}
      */
     public function load(array $configs, ContainerBuilder $container): void
     {